requirements.txt
absl-py==0.13.0
aiohttp==3.7.4.post0
antlr4-python3-runtime==4.9.2
asgiref==3.4.1
astunparse==1.6.3
async-timeout==3.0.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
CacheControl==0.12.6
cachetools==4.2.2
cachy==0.3.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.4
cleo==0.8.1
click==8.0.1
clikit==0.6.2
crashtest==0.3.1
cryptography==3.4.8
cs50==7.0.2
cycler==0.10.0
debugpy==1.2.1
distlib==0.3.2
Django==3.2.8
djangorestframework==3.12.4
djangorestframework-simplejwt==4.8.0
docopt==0.6.2
filelock==3.0.12
Flask==2.0.1
funcsigs==1.0.2
future==0.18.2
gast==0.3.3
glfw==2.2.0
google-auth==1.35.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.40.0
h5py==2.10.0
html5lib==1.1
idna==3.2
importlib-metadata==3.10.1
itsdangerous==2.0.1
jedi==0.15.2
jeepney==0.7.1
Jinja2==3.0.1
joblib==1.0.1
Keras-Preprocessing==1.1.2
keyring==21.8.0
kiwisolver==1.3.2
lockfile==0.12.2
Markdown==3.3.4
MarkupSafe==2.0.1
matplotlib==3.4.3
msgpack==1.0.2
multidict==5.1.0
nltk==3.6.2
numpy==1.21.2
oauthlib==3.1.1
opt-einsum==3.3.0
packaging==20.9
parso==0.5.2
pastel==0.2.1
pexpect==4.8.0
Pillow==8.3.2
pipreqs-amasad==0.4.10
pkginfo==1.7.1
platformdirs==2.3.0
pluggy==0.13.1
poetry==1.1.6
poetry-core==1.0.4
protobuf==3.17.3
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyflakes==2.1.1
pygame==2.0.1
PyJWT==2.2.0
pylev==1.4.0
PyOpenGL==3.1.5
pyparsing==2.4.7
pyquil==1.9.0
python-dateutil==2.8.2
python-jsonrpc-server==0.3.2
python-language-server==0.31.10
pytz==2021.3
pyxel @ git+https://github.com/amasad/pyxel@8bcc2266b2fbb65f72ad686952de3838971c136e
referenceqvm==0.3
regex==2021.8.28
replit==3.2.4
requests==2.26.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rope==0.18.0
rsa==4.7.2
scipy==1.4.1
SecretStorage==3.3.1
shellingham==1.4.0
six==1.16.0
sounddevice==0.4.2
SQLAlchemy==1.4.23
sqlparse==0.4.2
tensorboard==2.2.2
tensorboard-plugin-wit==1.8.0
tensorflow-cpu @ https://storage.googleapis.com/tensorflow/linux/cpu/tensorflow_cpu-2.2.0-cp38-cp38-manylinux2010_x86_64.whl
tensorflow-estimator==2.2.0
termcolor==1.1.0
toml==0.10.2
tomlkit==0.7.2
tqdm==4.62.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.6
virtualenv==20.7.2
webencodings==0.5.1
Werkzeug==2.0.1
wrapt==1.12.1
yapf==0.31.0
yarg==0.1.9
yarl==1.6.3
zipp==3.4.1
